leshop-crawler/requirements.txt
requests
wolframalpha
google-cloud-firestore